# backend/api/enhanced.py
# Phase 3: Enhancements - Enhanced API Endpoints

from fastapi import APIRouter, HTTPException, Depends
from pydantic import BaseModel, Field
from typing import List, Optional, Dict, Any
from datetime import datetime
import logging
from ..middleware.security import SecurityMiddleware

logger = logging.getLogger(__name__)

router = APIRouter(prefix="/enhanced", tags=["enhanced"])

# Security middleware for API validation
security_middleware = SecurityMiddleware()

# Request/Response models
class ProjectCreate(BaseModel):
    name: str = Field(..., min_length=1, max_length=100)
    description: Optional[str] = None
    settings: Dict[str, Any] = {}

class ProjectUpdate(BaseModel):
    name: Optional[str] = Field(None, min_length=1, max_length=100)
    description: Optional[str] = None
    settings: Optional[Dict[str, Any]] = None

class ExportConfig(BaseModel):
    format: str = Field(..., pattern="^png|jpeg|mp3|wav|json$")
    quality: int = Field(90, ge=1, le=100)
    include_metadata: bool = True

class BatchGenerateRequest(BaseModel):
    prompts: List[str] = Field(..., min_items=1, max_items=10)
    duration: float = Field(5.0, ge=1.0, le=60.0)

class BatchGenerateResponse(BaseModel):
    results: List[Dict[str, Any]]
    total: int
    successful: int
    failed: int

# Mock database for projects
projects_db: Dict[str, Dict[str, Any]] = {}

@router.get("/projects", response_model=List[Dict[str, Any]])
async def get_projects():
    """Get all projects"""
    return list(projects_db.values())

@router.get("/projects/{project_id}", response_model=Dict[str, Any])
async def get_project(project_id: str):
    """Get specific project details"""
    if project_id not in projects_db:
        raise HTTPException(status_code=404, detail="Project not found")
    return projects_db[project_id]

@router.post("/projects", response_model=Dict[str, Any], status_code=201)
async def create_project(project: ProjectCreate):
    """Create a new project"""
    project_id = f"proj_{datetime.utcnow().strftime('%Y%m%d%H%M%S')}"
    
    projects_db[project_id] = {
        "id": project_id,
        "name": project.name,
        "description": project.description,
        "settings": project.settings,
        "created_at": datetime.utcnow().isoformat(),
        "updated_at": datetime.utcnow().isoformat()
    }
    
    logger.info(f"Created project: {project_id}")
    return projects_db[project_id]

@router.put("/projects/{project_id}", response_model=Dict[str, Any])
async def update_project(project_id: str, project: ProjectUpdate):
    """Update existing project"""
    if project_id not in projects_db:
        raise HTTPException(status_code=404, detail="Project not found")
    
    if project.name:
        projects_db[project_id]["name"] = project.name
    if project.description is not None:
        projects_db[project_id]["description"] = project.description
    if project.settings is not None:
        projects_db[project_id]["settings"] = project.settings
    
    projects_db[project_id]["updated_at"] = datetime.utcnow().isoformat()
    
    logger.info(f"Updated project: {project_id}")
    return projects_db[project_id]

@router.delete("/projects/{project_id}", status_code=204)
async def delete_project(project_id: str):
    """Delete a project"""
    if project_id not in projects_db:
        raise HTTPException(status_code=404, detail="Project not found")
    
    del projects_db[project_id]
    logger.info(f"Deleted project: {project_id}")
    return None

@router.post("/export", response_model=Dict[str, Any])
async def export_project(project_id: str, config: ExportConfig):
    """Export project with specified format and quality"""
    if project_id not in projects_db:
        raise HTTPException(status_code=404, detail="Project not found")
    
    project = projects_db[project_id]
    
    # Mock export operation
    export_result = {
        "project_id": project_id,
        "format": config.format,
        "quality": config.quality,
        "metadata": project.get("settings", {}).get("metadata", {}),
        "exported_at": datetime.utcnow().isoformat()
    }
    
    logger.info(f"Exported project {project_id} as {config.format}")
    return export_result

@router.post("/batch-generate", response_model=BatchGenerateResponse)
async def batch_generate(request: BatchGenerateRequest):
    """Generate multiple items in batch for performance"""
    results = []
    successful = 0
    failed = 0
    
    for prompt in request.prompts:
        try:
            # Mock generation
            result = {
                "prompt": prompt,
                "status": "completed",
                "id": f"gen_{datetime.utcnow().strftime('%Y%m%d%H%M%S%f')}"
            }
            results.append(result)
            successful += 1
        except Exception as e:
            logger.error(f"Batch generation failed for prompt '{prompt}': {e}")
            results.append({
                "prompt": prompt,
                "status": "failed",
                "error": str(e)
            })
            failed += 1
    
    return BatchGenerateResponse(
        results=results,
        total=len(request.prompts),
        successful=successful,
        failed=failed
    )

@router.get("/health/detailed")
async def get_detailed_health():
    """Get detailed system health information"""
    return {
        "status": "healthy",
        "timestamp": datetime.utcnow().isoformat(),
        "services": {
            "backend": {"status": "operational", "version": "1.0.0"},
            "database": {"status": "connected", "type": "sqlite"},
            "cache": {"status": "connected", "type": "redis"},
            "ai_model": {"status": "ready", "model": "StableDiffusion-XL"}
        },
        "performance": {
            "average_response_time": 0.045,  # seconds
            "peak_memory_usage": 2.8,  # GB
            "gpu_utilization": 72
        }
    }