# backend/docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Unified Editing Studio API
  version: 1.0.0
  description: |
    Complete API for AI-powered sprite art and music generation.
    Supports image generation, audio generation, project management,
    and batch operations with real-time collaboration.
  contact:
    name: Unified Editing Studio Team
    email: support@unifiedediting.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://studio.unifiedediting.com
    description: Production server
...
paths:
...
  /api/v1/enhanced/projects:
    get:
      tags: [enhanced]
      summary: List all projects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
    post:
      tags: [enhanced]
      summary: Create new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Project created
...
  /api/v1/enhanced/batch-generate:
    post:
      tags: [enhanced]
      summary: Generate multiple items in batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompts:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Batch generation results
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
...